directory structure, speeded up --- - by calling only `addEvent . once - use `Promise.resolve()` speed up run cb ---
